refactor(object): apply review fixes
